removing local_project_dir and allowing a dir not to exist

+++ buildstockbatch/eagle.py
@@ -61,7 +61,6 @@ class EagleBatch(BuildStockBatchBase):
     min_sims_per_job = 36 * 2
 
     local_scratch = pathlib.Path(os.environ.get("LOCAL_SCRATCH", "/tmp/scratch"))
-    local_project_dir = local_scratch / "project"
     local_buildstock_dir = local_scratch / "buildstock"
     local_weather_dir = local_scratch / "weather"
     local_output_dir = local_scratch / "output"
@@ -314,9 +313,7 @@ class EagleBatch(BuildStockBatchBase):
 
         # Remove local scratch files
         dirs_to_remove = [
-            self.local_project_dir,
             self.local_buildstock_dir,
-            self.local_weather_dir,
             self.local_weather_dir,
             self.local_output_dir,
             self.local_housing_characteristics_dir,
@@ -325,7 +322,10 @@ class EagleBatch(BuildStockBatchBase):
         logger.info(f"Cleaning up {self.local_scratch}")
         for dir in dirs_to_remove:
             logger.debug(f"Removing {dir}")
-            shutil.rmtree(dir)
+            if dir.exists():
+                shutil.rmtree(dir)
+            else:
+                logger.warning(f"Directory does not exist {dir}")
         logger.debug(f"Removing {self.local_singularity_img}")
         os.remove(self.local_singularity_img)
